requirements.txt
Django==5.1.4
django-environ==0.11.2
google-generativeai==0.8.3
